revert switch to generic fixOptional
